feat: delete member service, schema='mo'/#11
